refactor: split analyzer to the main class and anutils; reorder methods in analyzer main class to make the file more logical to read for humans
